AEROGEAR-8027 - audit log to print client info coming with extensions
